Move tests that use the SEM manager into a separate subproject
