+++ galasa-internal-inttests-parent/dev.galasa.internal.inttests/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Version: 0.0.1-SNAPSHOT
+Bundle-Name: dev.galasa.internal.inttests
+Import-Package: *
