+++ galasa-internal-inttests-parent/dev.galasa.internal.inttests/src/main/resources/zosPrograms/PRGABEND.cbl
@@ -1,0 +1,10 @@
+       IDENTIFICATION DIVISION.
+       PROGRAM-ID. PRGABEND.
+       ENVIRONMENT DIVISION.
+       CONFIGURATION SECTION.
+       DATA DIVISION.
+       WORKING-STORAGE SECTION.
+       PROCEDURE DIVISION.
+           EXEC CICS ABEND
+           END-EXEC.
+           EXEC CICS RETURN END-EXEC.
