+++ galasa-internal-inttests-parent/gradle.properties
@@ -1,0 +1,3 @@
+sourceMaven=https://development.galasa.dev/main/maven-repo/obr
+internalMaven=https://repo.maven.apache.org/maven2/
+targetMaven=https://repo.maven.apache.org/maven2/
