+++ galasa-inttests-parent/gradle.properties
@@ -1,6 +1,6 @@
 galasaGroup=dev.galasa
 galasaName=galasa
-galasaVersion=0.36.0
+galasaVersion=0.38.0
 galasaSourceCompatibility=11
 galasaTargetCompatibility=11
